IN-5 fix: rename non component to camelCase, change children to render in storybook
