+++ src/shared/ui/Select/SelectItem.tsx
@@ -2,7 +2,7 @@ import * as React from "react";
 import {
   RadixSelectItem as Item,
   RadixSelectItemText as ItemText,
-} from "./SelectPrimitive";
+} from "src/shared/ui/Select/selectPrimitive";
 import { cn } from "@/features/utils/cn";
 
 const SelectItem = React.forwardRef<
